feat(ico): agregado ico al PSEmodel

+++ EpaycoSdk/Models/Bank/SplitReceivers.cs
@@ -3,10 +3,10 @@ namespace EpaycoSdk.Models.Bank
     public class SplitReceivers
     {
         public string id { get; set; }
-        public int total { get; set; }
-        public int iva { get; set; }
-        public int ico { get; set; }
-        public int base_iva { get; set; }
-        public int fee { get; set; }
+        public decimal total { get; set; }
+        public decimal iva { get; set; }
+        public decimal ico { get; set; }
+        public decimal base_iva { get; set; }
+        public decimal fee { get; set; }
     }
 }
